// bot/src/generated_schema/user_data/servers.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: hkXl/FJ3mZF1tsHtjZPJGYZAuAcGl0mNWZ3ZkvU3OR15HBid6U/djAXSHQdSe54+8UTIwkKBSdZ040jSPSva2Q==
 */

/* eslint-disable */
// tslint:disable

interface Servers {
  charseps: (Array<string | null>) | null
  charsigns: (Array<string | null>) | null
  codex: (number) | null
  ephemeral: (boolean) | null
  freeze_on_timeout: (boolean) | null
  inline: (boolean) | null
  server_id: string & {readonly __brand?: 'servers_server_id'}
}
export default Servers;

interface Servers_InsertParameters {
  charseps?: (Array<string | null>) | null
  charsigns?: (Array<string | null>) | null
  codex?: (number) | null
  ephemeral?: (boolean) | null
  freeze_on_timeout?: (boolean) | null
  inline?: (boolean) | null
  server_id: string & {readonly __brand?: 'servers_server_id'}
}
export type {Servers_InsertParameters}